+++ go.mod
@@ -4,4 +4,7 @@ go 1.18
 
 require github.com/essentialkaos/ek/v12 v12.111.1
 
-require golang.org/x/sys v0.18.0 // indirect
+require (
+	github.com/essentialkaos/depsy v1.1.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
+)
